Split Makefile into common and repo specific sections
* Other repos with their own docker build targets can inlcude common.mk
without overwriting their own docker targets.
* Normally we could achieve a similar result by placing wildcard symbols
(%) in the target names. However, because there is already a
sub-directory named 'docker', placing a wildcard in the 'docke%' target
name will cause the target to always be secondary to the directory
(regardless of .PHONY target) and therefore never meet constraints for
being rebuilt.