oauth: Add terminal redirect URL option
Add a `--redirect-url` string option to the boostrap command that allows
the terminal redirect in the OAuth flow to be configured. This is useful
if you want to display a more interesting page than the locally hosted
basebones "Success" page.

If the user configures step-cli using the `ssh config` command and
passes the `--team` option, set the redirect url to:

    https://smallstep.com/app/teams/sso/success

so that the flow ends on a page that instructs users on what to do next.