certificates: create: only force default SAN for leaf certs
Forcing this for root CA and intermediate CA certs was resulting
in the subject (typically a company name) being interpreted as
a dnsName SAN.  This produced an invalid certificate.  It is also
highly unusual to have a root CA with a SAN at all as seen in
censys so it's best to leave it out.

Fixes #263