add support for TLS Server Name Indication
Added the ability use TLS Server Name Indication (SNI) to specify
which certificate the server should send. This can be useful when
the server supports multiple certificates or when the DNS doesn't
match the IP address (e.g. in development scenarios) but we still
want certificate validation to be performed.

This option was added to all commands that support a remote fetch
of a peer certificate. These commands includes:
  - `fingerprint`
  - `inspect`
  - `lint`
  - `verify`