Clarify the use of the --redirect-url flag.
This change changes the usage of the --redirect-url. This flag is
used in the `step oauth` command and can also be used in the
`step bootstrap` command to write it in the defaults.json file.

This flag won't have any effect on oob flows, but we cannot fail
if both are used because it will cause errors on specific configurations
where errors are not expected.

Fixes #510