Allow to specify a fixed redirect_uri.
In some environments like containers it might be difficult to
redirect to 127.0.0.1 and we might need to listen in 0.0.0.0, but
at the same time we cannot use http://0.0.0.0:10000 as the redirect_uri.
This parameters allows to override the redirect_uri in a loopback
authorization.

The name is set to --listen-url instead of --redirect-uri to avoid
confusion with --redirect-url. And generally --listen and --listen-url
will be used together.