Add emoji format to fingerprint commands
Add an additional `emoji` format for the `step certificate fingerprint`
that encodes the certificate fingerprint as an emoji string.

Add the `--format` flag to the `step ssh fingerprint` command that
controls the encoding of the fingerprint. The default behavior of the
command is unchanged. The additional encodings are the same as for
`step certificate fingerprint`.

Add teh `--format` flag to the `step crypto key fingerprint` command that
controls the encoding of the fingerprint. The default behavior of the
command is unchanged. For `sha1` hex encoding is used, for `sha256`
base64 encoding is used. The additional encodings are the same as for
`step certificate fingerprint`.

fixes #570