Change authority and profile to default to context name when set
Before this change the authority and profile would default to
the first DNS name, potentially resulting in multiple CAs using
the same directories and configuration or configurations being
overwritten.

After this change the name of the context will be used as the
authority and profile name by default. They can still be overridden
using the `--authority` and `--profile` flags. When no context
name is provided, we still default to the first DNS name provided.

Fixes #606