Add `--remote-administration` flag to enable Admin API on `ca init`
This partially addresses https://github.com/smallstep/cli/issues/577,
but using `--remote-administration` as the flag name. It'll result
in the Admin API being enabled as well as a default provisioner and
corresponding admin credential being generated.

Currently this requires a DB to be configured too, as the logic
for creation of the new PKI stores the new provisioner and admin
in the DB at all times when the Admin API gets enabled.

The behavior this flag introduces might become the default at
some point.