Add `--admin-subject` to `ca init`
This allows the first super admin subject to be set when a new
CA is initialized. Usage is effectively limited to when remote
management is in use, because the super admin subject will only
be created when provisioners are stored in the database and the
admin API is enabled. It also doesn't work when the CA performs
automatic provisioner migration, as there's no nice way to pass
this information at the moment.

The flag can currently not be used with `--helm`.