Fix error when using `--insecure` and creating RSA keys <2048 bits
Before this commit, the CLI would allow specifying RSA key sizes
smaller than 2048 when also setting `--insecure`, but creation of
the key would still fail with `the size of the RSA key should be
at least 2048 bits`, originating form the `keyutil.generateRSAKey`
function.

This commit enables the insecure mode for the `keyutil` package
when `--insecure` is provided, allowing short RSA keys to be created.

While there are existing tests for checking (in)compatible CLI
flags that involve short RSA keys, the command didn't seem to be
actually executed as part of the test. That's part of the reason
why this bug wasn't found earlier, I think.