Pass key to be signed as signer for the CSR instead of signing key
Before this commit the CSR (used internally to prepare the final
certificate to be signed) was signed using the key that would sign
the final certificate instead of by the key to be signed. This
commit passes in `priv` instead of `signer`.

This could lead to non-backwards compatible issues, but I think
those shouldn't happen often.