Fix leaf certificate `*rsa.PublicKey` example
As reported in https://github.com/smallstep/cli/issues/1051, the
example template printed in the help has `*sa.PublicKey` instead of
`*rsa.PublicKey`. This is likely due to the help text preprocessor.

This commit adds an underscore to `*_rsa.PublicKey`, which tricks
the help text preprocessor into simply printing `*rsa.PublicKey`,
ignoring it to be processed as Markdown.

A better fix might be do change the help text processor, but there
are no tests, and it's used widely, so it's a riskier method for
not too much gain.