feat: add: --password-file option for step crypto jwe decrypt
The `encryptedKey` generated by `--type=jwk` in `step ca provisioner add` is in PBES2 format.
This is hard-coded and difficult to change.
Therefore, in the implementation before this commit, after the key to decrypt is entered in the standard input, the executor must interactively enter the password specified when generated in the standard input.
The `--key` option and others cannot be used at this time because of formatting problems.
However, trying to do this programmatically is difficult, as the format of `ui.PromptPassword` is too specific for manual execution, or reading stderr output, etc., and trying to add additional standard input passwords in a programmatic and automatic way is very It was very difficult.
Therefore, I added the `--password-file` command line argument to provide an alternative that does not require an interactive operation.
Since there was no problem reading the password file when generating it, I decided that there was no harm in reading the password file when decrypting it, and it was a natural implementation.