Use scopes from provider for OIDC token generation if present
This allows the provider to specify custom scopes that should be used
when generating OIDC tokens. This is useful when, for example, when
using the OIDC provider with Dex and Github in that it allows the
preferred_username or federated_claims.user_id fields to be returned
with the token for use in SSH certificate templates (ie. to identify a
Github user via extensions:login@github.com).

Note that this feature depends on
https://github.com/smallstep/certificates/pull/1796 being merged to
provide the `Scopes` field on `providers.OIDC`.