Add support for local GoReleaser builds with `goreleaser-local` target
The new `goreleaser-local` target allows building CLI releases that
are very similar to the ones that we release as part of our GitHub
Actions release process. The target skips the hooks that trigger
binaries to be pushed to GCP, allowing the release process to happen
locally. Otherwise it's the same as the existing `goreleaser` target,
which is intended for local snapshot builds of the CLI binary.

This PR also allows the GoReleaser process to be performed with
either its Pro or OSS version.