Use `envOrDefault` for `GPG_PRIVATE_KEY_FILE` environment variable
By using `envOrDefault` an issue is prevented in evaluating the
template before a release is created. By setting the default value
to `ENV_VAR_GPG_PRIVATE_KEY_FILE_NOT_SET`, this value is then shown
in the error message at signing time.

The reason for changing this is so that it becomes possible to
create local, unsigned releases when creating a release as follows:

 `goreleaser release --snapshot --clean --skip sign,after`